+++ src/app/app.component.html
@@ -4,18 +4,18 @@
   <div class="container-fluid">
     <div class="row">
       <div class="col-sm">
-        <div class="pc-chart-container">
-          <pc-bar-chart></pc-bar-chart>
+        <div class="pcac-chart-container">
+          <pcac-bar-chart></pcac-bar-chart>
         </div>
       </div>
       <div class="col-sm">
-        <div class="pc-chart-container">
-          <pc-table [config]="tableConfig"></pc-table>
+        <div class="pcac-chart-container">
+          <pcac-table [config]="tableConfig"></pcac-table>
         </div>
       </div>
       <div class="col-sm">
-        <div class="pc-chart-container">
-          <pc-line-area-chart></pc-line-area-chart>
+        <div class="pcac-chart-container">
+          <pcac-line-area-chart></pcac-line-area-chart>
         </div>
       </div>
     </div>
