+++ src/app/app.component.scss
@@ -1,6 +1,6 @@
 @import "../imports";
 
-.pc-chart-container {
+.pcac-chart-container {
   padding: 1rem;
   background: color("white");
   box-shadow: 10px 0px 2rem -12px rgba(90, 87, 87, 0.75);
