+++ src/app/pcac/header/header.component.html
@@ -1,0 +1,3 @@
+<div class="clearfix">
+  <div class="float-left">{{config.title}}</div>
+</div>
