+++ src/app/pcac/header/header.component.ts
@@ -1,11 +1,13 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
+import { IHeaderConfig } from './header.model';
 
 @Component({
-  selector: 'pc-header',
+  selector: 'pcac-header',
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.scss']
 })
 export class HeaderComponent implements OnInit {
+  @Input() config: IHeaderConfig;
 
   constructor() { }
