+++ src/app/pcac/line-area-chart/line-area-chart.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'pc-line-area-chart',
+  selector: 'pcac-line-area-chart',
   templateUrl: './line-area-chart.component.html',
   styleUrls: ['./line-area-chart.component.scss']
 })
