+++ src/app/pcac/pcac.module.ts
@@ -3,19 +3,18 @@ import { AppComponent } from '../app.component';
 import { BarChartComponent } from './bar-chart/bar-chart.component';
 import { TableComponent } from './table/table.component';
 import { LineAreaChartComponent } from './line-area-chart/line-area-chart.component';
-import { CommonModule } from '@angular/common';
-import { SharedModule } from '../shared/header.module';
+import { BrowserModule } from '@angular/platform-browser';
+import { HeaderComponent } from './header/header.component';
 
+// TODO: Why do we need to add AppComponent to this module?
 @NgModule({
-  imports: [
-    CommonModule,
-    SharedModule
-  ],
+  imports: [BrowserModule],
   declarations: [
     AppComponent,
     BarChartComponent,
     TableComponent,
-    LineAreaChartComponent
+    LineAreaChartComponent,
+    HeaderComponent
   ]
 })
-export class ChartModule { }
+export class PioneerCodeAngularChartModule { }
