+++ src/app/pcac/table/table.component.html
@@ -1,4 +1,4 @@
-<pc-header [headerConfig]=""></pc-header>
+<pcac-header></pcac-header>
 <table class="table">
   <thead>
     <tr>
