+++ src/app/pcac/table/table.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, Input } from '@angular/core';
 import { ITableConfig } from './table.model';
 
 @Component({
-  selector: 'pc-table',
+  selector: 'pcac-table',
   templateUrl: './table.component.html',
   styleUrls: ['./table.component.scss']
 })
