+++ src/app/pcac/table/table.model.ts
@@ -1,5 +1,5 @@
 import { IData } from '../data.model';
-import { IHeaderConfig } from '../../shared/header/header.model';
+import { IHeaderConfig } from '../header/header.model';
 
 export interface ITableConfig {
   headerConfig: IHeaderConfig;
