--- src/app/shared/header.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { HeaderComponent } from './header/header.component';
-
-@NgModule({
-  imports: [
-    CommonModule
-  ],
-  declarations: [
-    HeaderComponent
-  ]
-})
-export class SharedModule { }
