--- src/app/shared/header/header.component.html
@@ -1,3 +1,0 @@
-<p>
-  header works!
-</p>
