+++ tslint.json
@@ -121,8 +121,7 @@
       true,
       "attribute",
       [
-        "app",
-        "pc"
+        "pcac"
       ],
       "camelCase"
     ],
@@ -130,8 +129,7 @@
       true,
       "element",
       [
-        "app",
-        "pc"
+        "pcac"
       ],
       "kebab-case"
     ],
